+++ src/constants.ts
@@ -51,16 +51,17 @@ export enum RelayPrefix {
   WEBRTC_SIGNALLING
 }
 
-export const RELAY_PAYLOAD_PREFIX = Uint8Array.from([0])
-export const RELAY_STATUS_PREFIX = Uint8Array.from([1])
-export const RELAY_WEBRTC_PREFIX = Uint8Array.from([2])
-export const RELAY_CONNECTION_STATUS_PREFIX = Uint8Array.from([3])
-export const VALID_PREFIXES = [
-  ...RELAY_PAYLOAD_PREFIX,
-  ...RELAY_STATUS_PREFIX,
-  ...RELAY_WEBRTC_PREFIX,
-  ...RELAY_CONNECTION_STATUS_PREFIX
-]
+export function isValidPrefix(prefix: RelayPrefix): boolean {
+  switch (prefix) {
+    case RelayPrefix.PAYLOAD:
+    case RelayPrefix.STATUS_MESSAGE:
+    case RelayPrefix.CONNECTION_STATUS:
+    case RelayPrefix.WEBRTC_SIGNALLING:
+      return true
+    default:
+      return false
+  }
+}
 
 export const RELAY = `/${NAME}/relay/${version}`
 export const DELIVERY = `/${NAME}/delivery/${version}`
